🚑 [Fix]: OptionGroup 디자인 시스템과 그룹핑 동기화(Option -> Field) - CheckBoxLabel, Label 컴포넌트 활용
